Crate hopr_transport_tag_allocator

Re-exports

pub use config::TagAllocatorConfig;

pub use errors::TagAllocatorError;

Modules

config

errors

Structs

AllocatedTag

A tag value that is automatically returned to its allocator on drop.

Enums

Usage

Identifies which component a partition belongs to.

Constants

TAG_RANGE_END

Upper bound of the application tag range (exclusive).

TAG_RANGE_SIZE

Total number of available application tags.

TAG_RANGE_START

Lower bound of the application tag range (exclusive end of ReservedTag range).

Traits

TagAllocator

Allocates unique tags from a fixed partition of the tag range.

Functions

create_allocators

Create one TagAllocator per partition from a contiguous tag range.

create_allocators_from_config

Create allocators from a TagAllocatorConfig.

Type Aliases

CreateAllocatorsResult

Result type returned by create_allocators.